requirements.txt
duckdb
editdistance
jellyfish
pandas
nlpaug
jaro-winkler
oracledb
